Split code into multiple modules
